Scripts and Datasets used in R and for tableau
Here are the scripts used to simulate the initial data and connect to the MongoDB bank. In addition, there is the script in which the data obtained by the voice assistant, the forecast data of the algorithm we use and bed data, and these data were imported from MongoDB, with the exception of the RDS file.